Refact validation model plugin test
Related to #559.
