--- examples/tests/hu.elte.txtuml.examples.validation/src/hu/elte/txtuml/examples/validation/InvalidClass.java
@@ -1,4 +1,0 @@
-package hu.elte.txtuml.examples.validation;
-public class InvalidClass {
-
-}
