--- examples/tests/hu.elte.txtuml.examples.validation/src/hu/elte/txtuml/examples/validation/InvalidFieldType.java
@@ -1,14 +1,0 @@
-package hu.elte.txtuml.examples.validation;
-import hu.elte.txtuml.api.model.ModelClass;
-import hu.elte.txtuml.examples.validation.helpers.A;
-import hu.elte.txtuml.examples.validation.helpers.ExternalClass;
-import hu.elte.txtuml.examples.validation.helpers.MyDataType;
-
-public class InvalidFieldType extends ModelClass {
-
-	Object x;
-	A a;
-	ExternalClass external;
-	MyDataType datatype;
-
-}
