+++ examples/tests/hu.elte.txtuml.examples.validation/src/hu/elte/txtuml/examples/validation/model/DataTypeInvalidFieldType.java
@@ -1,8 +1,8 @@
-package hu.elte.txtuml.examples.validation;
+package hu.elte.txtuml.examples.validation.model;
 
 import hu.elte.txtuml.api.model.DataType;
-import hu.elte.txtuml.examples.validation.helpers.A;
-import hu.elte.txtuml.examples.validation.helpers.MyDataType;
+import hu.elte.txtuml.examples.validation.model.helpers.A;
+import hu.elte.txtuml.examples.validation.model.helpers.MyDataType;
 
 public class DataTypeInvalidFieldType extends DataType {
