+++ examples/tests/hu.elte.txtuml.examples.validation/src/hu/elte/txtuml/examples/validation/model/InvalidClass.java
@@ -1,0 +1,4 @@
+package hu.elte.txtuml.examples.validation.model;
+public class InvalidClass {
+
+}
