+++ examples/tests/hu.elte.txtuml.examples.validation/src/hu/elte/txtuml/examples/validation/model/InvalidFieldType.java
@@ -1,0 +1,14 @@
+package hu.elte.txtuml.examples.validation.model;
+import hu.elte.txtuml.api.model.ModelClass;
+import hu.elte.txtuml.examples.validation.model.helpers.A;
+import hu.elte.txtuml.examples.validation.model.helpers.ExternalClass;
+import hu.elte.txtuml.examples.validation.model.helpers.MyDataType;
+
+public class InvalidFieldType extends ModelClass {
+
+	Object x;
+	A a;
+	ExternalClass external;
+	MyDataType datatype;
+
+}
