+++ examples/tests/hu.elte.txtuml.examples.validation/src/hu/elte/txtuml/examples/validation/model/InvalidParameterType.java
@@ -1,8 +1,8 @@
-package hu.elte.txtuml.examples.validation;
+package hu.elte.txtuml.examples.validation.model;
 import hu.elte.txtuml.api.model.ModelClass;
-import hu.elte.txtuml.examples.validation.helpers.A;
-import hu.elte.txtuml.examples.validation.helpers.ExternalClass;
-import hu.elte.txtuml.examples.validation.helpers.MyDataType;
+import hu.elte.txtuml.examples.validation.model.helpers.A;
+import hu.elte.txtuml.examples.validation.model.helpers.ExternalClass;
+import hu.elte.txtuml.examples.validation.model.helpers.MyDataType;
 
 public class InvalidParameterType extends ModelClass {
