+++ examples/tests/hu.elte.txtuml.examples.validation/src/hu/elte/txtuml/examples/validation/model/InvalidTypeInSignal.java
@@ -1,6 +1,6 @@
-package hu.elte.txtuml.examples.validation;
+package hu.elte.txtuml.examples.validation.model;
 import hu.elte.txtuml.api.model.Signal;
-import hu.elte.txtuml.examples.validation.helpers.A;
+import hu.elte.txtuml.examples.validation.model.helpers.A;
 
 public class InvalidTypeInSignal extends Signal {
 	int x;
