+++ examples/tests/hu.elte.txtuml.examples.validation/src/hu/elte/txtuml/examples/validation/model/TransitionOutsideOfClass.java
@@ -1,4 +1,4 @@
-package hu.elte.txtuml.examples.validation;
+package hu.elte.txtuml.examples.validation.model;
 
 import hu.elte.txtuml.api.model.From;
 import hu.elte.txtuml.api.model.ModelClass;
